Move basic exerice to own package
