+++ app/src/main/java/org/simonolander/lambda/content/exercise/basic/Identity.kt
@@ -1,4 +1,4 @@
-package org.simonolander.lambda.content.exercise
+package org.simonolander.lambda.content.exercise.basic
 
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.withStyle
@@ -6,45 +6,51 @@ import org.simonolander.lambda.domain.DialogBuilder
 import org.simonolander.lambda.domain.Exercise
 import org.simonolander.lambda.domain.TestCase
 import org.simonolander.lambda.engine.Application
-import org.simonolander.lambda.engine.Function
 import org.simonolander.lambda.engine.Identifier
 import org.simonolander.lambda.ui.theme.codeStyle
 
-val constantExercise = run {
-    val name = "A constant function"
-    val functionName = "const_a"
+val identityExercise = run {
+    val name = "Identity"
+    val functionName = "id"
+
     val description = buildAnnotatedString {
         append("Design a function ")
         withStyle(codeStyle) { append(functionName) }
-        append(", that for every input ")
+        append(", that given any single input ")
         withStyle(codeStyle) { append("x") }
-        append(" produces the output ")
-        withStyle(codeStyle) { append("a") }
+        append(" produces the same output ")
+        withStyle(codeStyle) { append("x") }
         append(". ")
         append("\n\n")
         append("For example, ")
-        withStyle(codeStyle) { append("$functionName x") }
+        withStyle(codeStyle) { append("$functionName a") }
         append(" should reduce to ")
         withStyle(codeStyle) { append("a") }
         append(".")
     }
+
     val testCases = listOf(
-        Identifier("x"),
-        Identifier("a"),
-        Function("x", Identifier("x")),
+        "a",
+        "value",
+        "⛄️",
     ).map { arg ->
         TestCase(
             input = Application(
                 function = Identifier(functionName),
-                argument = arg,
+                argument = Identifier(arg),
             ),
-            output = Identifier("a"),
+            output = Identifier(arg),
         )
     }
 
     val dialog = DialogBuilder()
-        .message("Constant functions always return the same thing, regardless of their inputs.")
-        .message("They can be very useful!")
+        .message("Hello!")
+        .message("Enough talking, it's time for some action!")
+        .message("Here, you need to enter an the expression that satisfies the given exercise.")
+        .message("In this instance, I'm asking you to design the identity function.")
+        .message("Write it in the text field called id.")
+        .message("If you don't have a λ on your keyboard, you can simply click the icon in the text field.")
+        .message("Once you're happy with your function, click go and we'll see if it works!")
         .build()
 
     Exercise(
